Get config from Octopolo
When calling Octopolo::Config.parse on every class, we end up loading
the config more then we need to when it should always be the same data.
Also, it makes it munch harder to stub because it requires that it loads
the data from multiple different locations and classes.  Doing this
centralizes it and makes it much easier to work with.