Updates to Octopolo::GitHub::PullRequestCreator.edit_body
Fixes working with the tempfile so the contents are there when editing
and returning the value.  Since we sending the file for editing in the
via another program, we need ruby to close the file before doing so,
then re-open it to read the user added content

Also fixes editing the file by using the system call.  This allows the
ruby process to maintain being the parent process, and when the user is
done editing the file, the foreground process is returned to the ruby
process to finish out it's work.